requirements/base.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
Django
django-allauth
django-cleanup
idna
oauthlib
Pillow
PyJWT
python-dateutil
python3-openid
requests
requests-oauthlib
six
sqlparse
toml
urllib3
python-decouple
pytz
